pyproject.toml
[tool.poetry]
name = "social_flask"
version = "0.1.0"
description = ""
authors = ["Jardel Baia <jardelbaiajr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
Flask-SQLAlchemy = "^2.5.1"
python-dotenv = "^0.18.0"
psycopg2-binary = "^2.9.1"
Flask-Migrate = "^3.0.1"
Flask-RESTful = "^0.3.9"
Flask-Bcrypt = "^0.7.1"
marshmallow = "^3.12.2"
Flask-JWT-Extended = "^4.2.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
